Move print_diff to cli/diff.rs
