+++ cli/fmt.rs
@@ -7,11 +7,10 @@
 //! the future it can be easily extended to provide
 //! the same functions as ops available in JS runtime.
 
-use crate::colors;
+use crate::diff::print_diff;
 use crate::fs::files_in_subtree;
 use crate::op_error::OpError;
 use deno_core::ErrBox;
-use difference::{Changeset, Difference};
 use dprint_plugin_typescript as dprint;
 use std::fs;
 use std::io::stdin;
@@ -55,109 +54,6 @@ pub async fn format(args: Vec<String>, check: bool) -> Result<(), ErrBox> {
   } else {
     format_source_files(config, target_files).await
   }
-}
-
-fn print_diff(file_path: &PathBuf, orig: &str, edit: &str) {
-  let line_number_width = edit.split('\n').count().to_string().chars().count();
-  let print_line_number = |line| {
-    print!(
-      "{:0width$}{} ",
-      line,
-      colors::gray("|".to_string()),
-      width = line_number_width
-    );
-  };
-  let print_add = || {
-    print!("{}", colors::green_bold("+".to_string()));
-  };
-  let print_add_text = |s: String| {
-    print!("{}", colors::green(s));
-  };
-  let print_add_text_highlited = |s: String| {
-    print!("{}", colors::white_on_green(s));
-  };
-  let print_rem = || {
-    print!("{}", colors::red_bold("-".to_string()));
-  };
-  let print_rem_text = |s: String| {
-    print!("{}", colors::red(s));
-  };
-
-  println!();
-  println!(
-    "{} {}:",
-    colors::bold("from".to_string()),
-    file_path.display().to_string()
-  );
-  let Changeset { diffs, .. } = Changeset::new(orig, edit, "\n");
-  let mut line = 1;
-  for i in 0..diffs.len() {
-    match diffs[i] {
-      Difference::Add(ref x) => {
-        match diffs[i - 1] {
-          Difference::Rem(ref y) => {
-            print_line_number(line);
-            print_add();
-            let Changeset { diffs, .. } = Changeset::new(y, x, "");
-            let mut inline = line;
-            for c in diffs {
-              match c {
-                Difference::Same(ref z) => {
-                  let split = z.split('\n').enumerate();
-                  for (i, s) in split {
-                    if i > 0 {
-                      inline += 1;
-                      println!();
-                      print_line_number(inline);
-                      print_add();
-                    }
-                    print_add_text(s.to_string());
-                  }
-                }
-                Difference::Add(ref z) => {
-                  let split = z.split('\n').enumerate();
-                  for (i, s) in split {
-                    if i > 0 {
-                      inline += 1;
-                      println!();
-                      print_line_number(inline);
-                      print_add();
-                    }
-                    print_add_text_highlited(s.to_string());
-                  }
-                }
-                _ => (),
-              }
-            }
-            println!()
-          }
-          _ => {
-            let split = x.split('\n').enumerate();
-            for (i, s) in split {
-              print_line_number(line + i);
-              print_add();
-              print_add_text(s.to_string());
-              println!()
-            }
-          }
-        };
-        line += 1 + x.matches('\n').count();
-      }
-      Difference::Rem(ref x) => {
-        let split = x.split('\n').enumerate();
-        for (i, s) in split {
-          print_line_number(line + i);
-          print_rem();
-          print_rem_text(s.to_string());
-          println!()
-        }
-      }
-      Difference::Same(ref x) => {
-        line += 1 + x.matches('\n').count();
-      }
-    }
-  }
-  println!();
 }
 
 async fn check_source_files(
